Implement Big Chungus a.k.a. fatMap
`fatMap` is like `flatMap` but provides a sliding window into the source
sequence, so you can see both behind and in front of you at each step.
This is similar in concept to the `index` parameter for the basic array
map callback, but avoids the need to iterate through an entire lazy
sequence in advance.