Refactor Chungus sliding window
It took me forever to work out all the math on this.  The problem is
that, besides being a sliding window, the focal point *within* the
window can vary as well, and since it's a circular buffer, it's not easy
to tell where it begins and ends without explicit variables to track
that.  It ended up being easier just to refactor the whole thing into
its own class.