Fix bug when Big Chungus is too fat
Another few hours of brain-wracking work later, but I think I FINALLY
got it this time!  It turned out that if the window size for `fatMap`
was bigger than the sequence being mapped over, the logic went totally
screwy and didn't return the correct values.  Now everything is finally
right with the world.

...at least until Big Chungus eats it.  :rabbit: :earth_americas: