Make fatMap selector return plain values
...instead of queryables.  This tends to be more useful in practice,
I've found, e.g. in implementing convolution filters.  Plus it gives it
a nice duality with flatMap: fat vs. flat, depending on which side of
the pipe (input or output) the array is on.

Not sure if a sliding-window flatMap would actually be useful in
practice too.  I haven't found any use cases yet, at least.